move H5Dataset, H5SpecReader and H5SpecWriter to h5_utils
